:recycle: :: [#222] StudentDomain / Project, Prize Response 재사용성
