+++ Projects/Domain/StudentDomain/Sources/DTO/Response/FetchStudentDetailByTeacherResponseDTO.swift
@@ -35,82 +35,11 @@ public struct FetchStudentDetailByTeacherResponseDTO: Decodable {
         }
         // swiftlint: enable nesting
     }
-
-    public struct ProjectResponseDTO: Decodable {
-        public let name: String
-        public let iconImageURL: String
-        public let previewImageURLs: [String]
-        public let description: String
-        public let links: [LinkResponseDTO]
-        public let techStacks: [String]
-        public let myActivity: String
-        public let inProgress: InProgressResponseDTO
-
-        enum CodingKeys: String, CodingKey {
-            case name
-            case iconImageURL = "icon"
-            case previewImageURLs = "previewImages"
-            case description
-            case links
-            case techStacks
-            case myActivity
-            case inProgress
-        }
-    }
-
-    public struct PrizeResponseDTO: Decodable {
-        public let name: String
-        public let type: String
-        public let date: String
-    }
-
-    public struct LinkResponseDTO: Decodable {
-        public let name: String
-        public let url: String
-    }
-
-    public struct InProgressResponseDTO: Decodable {
-        public let start: String
-        public let end: String?
-    }
 }
 
 public extension FetchStudentDetailByTeacherResponseDTO.LanguageCertificateResponseDTO {
     func toDomain() -> LanguageCertificateEntity {
         LanguageCertificateEntity(name: name, score: score)
-    }
-}
-
-public extension FetchStudentDetailByTeacherResponseDTO.ProjectResponseDTO {
-    func toDomain() -> ProjectEntity {
-        ProjectEntity(
-            name: name,
-            iconImageURL: iconImageURL,
-            previewImageURLs: previewImageURLs,
-            description: description,
-            links: links.map { $0.toDomain() },
-            techStacks: techStacks,
-            myActivity: myActivity,
-            inProgress: inProgress.toDomain()
-        )
-    }
-}
-
-public extension FetchStudentDetailByTeacherResponseDTO.LinkResponseDTO {
-    func toDomain() -> ProjectEntity.LinkEntity {
-        ProjectEntity.LinkEntity(name: name, url: url)
-    }
-}
-
-public extension FetchStudentDetailByTeacherResponseDTO.InProgressResponseDTO {
-    func toDomain() -> ProjectEntity.InProgressEntity {
-        .init(start: start, end: end)
-    }
-}
-
-public extension FetchStudentDetailByTeacherResponseDTO.PrizeResponseDTO {
-    func toDomain() -> PrizeEntity {
-        PrizeEntity(name: name, type: type, date: date)
     }
 }
